backend/pyproject.toml
[project]
name = "TriviaQuizGame"
version = "0.1.0"
description = "A turn based quiz"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.2",
]
